+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 BUG FIXES:
 
 - **ontap_flexcache_resource**, **ontap_flexcache_data_source**, **ontap_flexcaches_data_source**: Fix use of use_tiered_aggregate in netapp-ontap_flexcache
+- **netapp-ontap_snapshot_policy_resource**: fixed issue with cluster scoped snapshot policy import ([#696](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/696))
 
 # 2.6.1 (2026-04-21)